requirements.txt
python-telegram-bot==20.8
requests==2.31.0